requirements.txt
torch
torchvision
pillow
matplotlib
opencv-python
numpy
scikit-image
plotly
grad-cam
streamlit
